+++ src/sphinx_substitution_extensions/__init__.py
@@ -8,10 +8,13 @@ import logging
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, ClassVar
 
+import docutils.nodes
 from docutils.parsers.rst import directives
 from docutils.parsers.rst.roles import code_role
 from docutils.parsers.rst.states import Inliner
+from sphinx import addnodes
 from sphinx.directives.code import CodeBlock
+from sphinx.roles import XRefRole
 
 from sphinx_substitution_extensions.extras import SubstitutionPrompt
 from sphinx_substitution_extensions.shared import (
@@ -19,9 +22,9 @@ from sphinx_substitution_extensions.shared import (
 )
 
 if TYPE_CHECKING:
-    import docutils.nodes
-    from docutils.nodes import Node, system_message
+    from docutils.nodes import Element, Node, system_message
     from sphinx.application import Sphinx
+    from sphinx.environment import BuildEnvironment
 
 LOGGER = logging.getLogger(__name__)
 
@@ -121,6 +124,41 @@ class SubstitutionCodeRole:
         return result_nodes, system_messages
 
 
+class SubstitutionXRefRole(XRefRole):
+    """Custom role for XRefs."""
+
+    def process_link(
+        self,
+        env: BuildEnvironment,
+        refnode: Element,
+        # We allow a boolean-typed positional argument as we are matching the
+        # method signature of the parent class.
+        has_explicit_title: bool,  # noqa: FBT001
+        title: str,
+        target: str,
+    ) -> tuple[str, str]:
+        """
+        Override parent method to replace placeholders with given variables.
+        """
+        document = self.inliner.document  # type: ignore[attr-defined]
+        assert isinstance(document, docutils.nodes.document)
+        for name, value in document.substitution_defs.items():
+            assert isinstance(name, str)
+            replacement = value.astext()
+            title = title.replace(f"|{name}|", replacement)
+            target = target.replace(f"|{name}|", replacement)
+
+        # Use the default implementation to process the link
+        # as it handles whitespace in target text.
+        return super().process_link(
+            env=env,
+            refnode=refnode,
+            has_explicit_title=has_explicit_title,
+            title=title,
+            target=target,
+        )
+
+
 def setup(app: Sphinx) -> dict[str, Any]:
     """
     Add the custom directives to Sphinx.
@@ -130,4 +168,8 @@ def setup(app: Sphinx) -> dict[str, Any]:
     app.setup_extension("sphinx-prompt")
     directives.register_directive("prompt", SubstitutionPrompt)
     app.add_role("substitution-code", SubstitutionCodeRole())
+    substitution_download_role = SubstitutionXRefRole(
+        nodeclass=addnodes.download_reference
+    )
+    app.add_role("substitution-download", substitution_download_role)
     return {"parallel_read_safe": True}